chore: move router director to app directory
