+++ lib/app/feature/authentication/auth/sign_in/sign_in_screen.dart
@@ -7,7 +7,7 @@ import '../../../../../core/ui/widget/elevated_button_widget.dart';
 import '../../../../../core/ui/widget/text_form_field_widget.dart';
 import '../../../../../core/ui/widget/widget_util.dart';
 import '../../../../../core/ui/widget/wrap_text_button.dart';
-import '../../../../../router/app_route.dart';
+import '../../../../router/app_route.dart';
 import '../../../../bloc/auth/auth_bloc.dart';
 import '../../../../common/util/app_strings.dart';
 import '../../../../common/util/key_strings.dart';
