+++ lib/app/feature/splash/splash_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../core/domain/manager/shared_prefs_manager.dart';
-import '../../../router/app_route.dart';
+import '../../router/app_route.dart';
 import '../../common/util/key_strings.dart';
 
 class SplashScreen extends StatelessWidget {
